pulled out config code into own file.
